Review fix - move new events to exising "private" namespace
