+++ ydb/core/tx/schemeshard/schemeshard__login.cpp
@@ -3,7 +3,6 @@
 #include <ydb/core/base/auth.h>
 #include <ydb/core/base/local_user_token.h>
 #include <ydb/core/protos/auth.pb.h>
-
 #include <ydb/library/login/login.h>
 #include <ydb/library/security/util.h>
 
@@ -66,14 +65,14 @@ struct TSchemeShard::TTxLogin : TSchemeShard::TRwTxBase {
         if (Self->LoginProvider.NeedVerifyHash(loginRequest, &Response, &passwordHash)) {
             ctx.Send(
                 Self->LoginHelper,
-                MakeHolder<TEvVerifyPassword>(loginRequest, Response, Request->Sender, passwordHash),
+                MakeHolder<TEvPrivate::TEvVerifyPassword>(loginRequest, Response, Request->Sender, passwordHash),
                 0,
                 Request->Cookie
             );
         } else {
             ctx.Send(
                 Self->SelfId(),
-                MakeHolder<TEvLoginFinalize>(loginRequest, Response, Request->Sender, "", /*needUpdateCache*/ false),
+                MakeHolder<TEvPrivate::TEvLoginFinalize>(loginRequest, Response, Request->Sender, "", /*needUpdateCache*/ false),
                 0,
                 Request->Cookie
             );
