+++ ydb/core/tx/schemeshard/ut_helpers/helpers.cpp
@@ -14,6 +14,7 @@
 #include <ydb/core/tx/data_events/events.h>
 #include <ydb/core/tx/data_events/payload_helper.h>
 #include <ydb/core/tx/schemeshard/schemeshard.h>
+#include <ydb/core/tx/schemeshard/schemeshard_private.h>
 #include <ydb/core/tx/sequenceproxy/sequenceproxy.h>
 #include <ydb/core/tx/tx_proxy/proxy.h>
 #include <ydb/core/util/pb.h>
@@ -2137,7 +2138,7 @@ namespace NSchemeShardUT_Private {
         const TString& passwordHash,
         const bool needUpdateCache
     ) {
-        const auto evLoginFinalize = new NKikimr::NSchemeShard::TEvLoginFinalize(
+        const auto evLoginFinalize = new NSchemeShard::TEvPrivate::TEvLoginFinalize(
             request, checkResult, runtime.AllocateEdgeActor(), passwordHash, needUpdateCache
         );
         AsyncSend(runtime, TTestTxConfig::SchemeShard, evLoginFinalize);
